Add files via upload
Three other students and me worked on this assignment for the Machine learning course of TiU. The goal was to develop a machine learning model that can predict the publication year of scientific articles based on metadata (abstract, publisher, editor, author, and entrytype). The evaluation metric was mean absolute error. We found that an ensemble model, which averages the predictions of ridge regression and stochastic gradient descent, gave us a sufficiently low MAE of 3.34. The hyperparameters were chosen after conducting several gridsearches. More information is available in the report.